Make gradients and shapes density independent.

+++ app/src/main/java/pl/karol202/paintplus/activity/ActivityPaintLayers.java
@@ -33,7 +33,7 @@ class ActivityPaintLayers implements View.OnClickListener
 		this.activity = activity;
 		
 		decorView = activity.getWindow().getDecorView();
-		sheetPanelSizePx = Utils.dpToPixels(activity.getDisplayMetrics(), SHEET_PANEL_SIZE_DP);
+		sheetPanelSizePx = (int) Utils.dpToPixels(activity.getDisplayMetrics(), SHEET_PANEL_SIZE_DP);
 	}
 	
 	void initLayers()
